drm/i915/preempt: Use GuC to preempt currently executing workload
Now that we're able to post-process the preemption event, let's actually
send it to GuC. To do this, we need to add a new work item to
preempt_client, and send a GuC action, rather than ringing the doorbell.
Preemption work item is just a HWSP write followed by an interrupt.
Because we want to avoid touching any of the inflight requests, we are
using separate preempt context.

Signed-off-by: Michał Winiarski <michal.winiarski@intel.com>