drm/i915/preempt: Handle preemption event in GuC tasklet
Preemption event is just ordered HWSP write, followed by an interrupt.
When preemption is in progress, we need to avoid submitting more work to
the HW. Once the preemption is completed, we know that the engine is
idle, so we can inspect the timeline and figure out which requests have
completed, and which need to be moved back to execution queue.

Signed-off-by: Michał Winiarski <michal.winiarski@intel.com>