drm/i915/preempt: Decide whether we need to preempt
After schedule, we need to figure out if we need to preempt any of the
currently executing requests, and for which engines we should attempt to
do it. To keep things simple, we're starting with only requesting
preemption for requests that have no dependencies between engines. We're
also always preempting if the scheduled request has priority above
threshold. This may result in preempting more important work, but since
it will be resubmitted right away we're still going to be functionally
correct.

Signed-off-by: Michał Winiarski <michal.winiarski@intel.com>