drm/i915/scheduler: Extract requests resubmission into a helper
We're going to use that for preemption soon. We can also leverage the
fact that resubmitted requests are likely to have the same priority, so
we can skip priolist lookup unless we observe a priority change.
To maintain the symmetry between request_in/request_out tracepoints,
let's also add a missing request_out tracepoint when resubmitting.

Signed-off-by: Michał Winiarski <michal.winiarski@intel.com>